move settings into conf file
