+++ lbrynet/conf.py
@@ -226,3 +226,11 @@ class Config(DefaultSettings):
     @property
     def UI_ADDRESS(self):
         return "http://%s:%i" % (DEFAULT_SETTINGS.API_INTERFACE, self.api_port)
+
+
+# TODO: don't load the configuration automatically. The configuration
+#       should be loaded at runtime, not at module import time. Module
+#       import should have no side-effects. This is also bad because
+#       it means that settings are read from the environment even for
+#       tests, which is rarely what you want to happen.
+settings = Config()
