+++ lbrynet/lbrynet_daemon/Daemon.py
@@ -26,7 +26,7 @@ from lbrynet import __version__ as lbrynet_version
 from lbryum.version import LBRYUM_VERSION as lbryum_version
 
 from lbrynet import __version__ as lbrynet_version
-from lbrynet import settings as lbrynet_settings
+from lbrynet.conf import settings as lbrynet_settings
 from lbrynet import analytics
 from lbrynet import reflector
 from lbrynet.metadata.Metadata import Metadata, verify_name_characters
